requirements.txt
pymupdf
llama-cpp-python
